+++ lib/tako/active_record_ext/singular_association.rb
@@ -1,0 +1,18 @@
+module Tako
+  module ActiveRecordExt
+    module SingularAssociation
+      def self.included(mod)
+        mod.extend(ShardedMethods)
+        mod.sharded_methods :reader,
+                            :writer,
+                            :create,
+                            :create!,
+                            :build
+      end
+    end
+  end
+end
+
+ActiveRecord::Associations::SingularAssociation.class_eval do
+  include Tako::ActiveRecordExt::SingularAssociation
+end
